Custom call logging plugin
