+++ auth-service/src/main/kotlin/uk/gov/communities/delta/auth/security/ADLdapLoginService.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.withContext
 import org.jetbrains.annotations.Blocking
 import org.slf4j.LoggerFactory
 import uk.gov.communities.delta.auth.config.LDAPConfig
-import uk.gov.communities.delta.auth.plugins.SpanFactory
+import uk.gov.communities.delta.auth.plugins.monitoring.SpanFactory
 import uk.gov.communities.delta.auth.repositories.LdapRepository
 import uk.gov.communities.delta.auth.repositories.LdapUser
 import javax.naming.AuthenticationException
